requirements.txt
python-dotenv==1.0.0
websockets==11.0.3
google-generativeai==0.3.1